Refactored add track panel, remote tracks take less space
